feat: Add support of Xcode27
