+++ .gitignore
@@ -6,3 +6,4 @@ test/assets
 coverage
 .nyc_output/
 package-lock.json*
+test/fixtures/*.app
