+++ lib/utils/constants.ts
@@ -1,0 +1,6 @@
+export const SAFARI_STARTUP_TIMEOUT_MS = 25 * 1000;
+export const MOBILE_SAFARI_BUNDLE_ID = 'com.apple.mobilesafari';
+export const SIMULATOR_APP_NAME = 'Simulator.app';
+export const DEVICE_HUB_APP_NAME = 'DeviceHub.app';
+export const MIN_SUPPORTED_XCODE_VERSION = 14;
+export const MIN_DEVICE_HUB_XCODE_VERSION = 27;
