+++ lib/utils/types.ts
@@ -1,0 +1,3 @@
+export interface SimulatorInfoOptions {
+  devicesSetPath?: string | null;
+}
