+++ test/unit/simulator-specs.ts
@@ -5,6 +5,7 @@ import sinon from 'sinon';
 import {devices} from './device-list';
 import {SimulatorXcode14} from '../../lib/simulator-xcode-14';
 import {SimulatorXcode15} from '../../lib/simulator-xcode-15';
+import {SimulatorXcode27} from '../../lib/simulator-xcode-27';
 import {use as chaiUse, expect} from 'chai';
 import chaiAsPromised from 'chai-as-promised';
 import * as xcodeModule from 'appium-xcode';
@@ -44,10 +45,16 @@ describe('simulator', function () {
     });
 
     const xcodeVersions: Array<
-      [number, number, string, typeof SimulatorXcode14 | typeof SimulatorXcode15]
+      [
+        number,
+        number,
+        string,
+        typeof SimulatorXcode14 | typeof SimulatorXcode15 | typeof SimulatorXcode27,
+      ]
     > = [
       [14, 0, '14.0.0', SimulatorXcode14],
       [15, 0, '15.0.0', SimulatorXcode15],
+      [27, 0, '27.0.0', SimulatorXcode27],
     ];
 
     for (const [major, minor, versionString, expectedXcodeClass] of xcodeVersions) {
